+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Unreleased
 
-  - …
+  - Add BaseNode.equals for deep-equality testing.
+
+    Nodes are deeply compared on a field by field basis. If possible, False is
+    returned early. When comparing attributes, tags and variants in
+    SelectExpressions, the order doesn't matter. By default, spans are not
+    taken into account.
 
 ## fluent 0.4.0 (June 13th, 2017)